blog delete by separate role

+++ src/app/modules/Blog/blog.services.ts
@@ -1,24 +1,69 @@
+import config from '../../config'
 import AppError from '../../error/AppError'
+import IUser from '../auth/user.interface'
 import { user } from '../auth/user.model'
 import IBlog from './blog.interface'
 import { blog } from './blog.model'
+import jwt, { JwtPayload } from 'jsonwebtoken'
 
 const createBlogIntroDB = async (payload: IBlog) => {
   const User = await user.findOne({ _id: payload.author })
   if (!User) {
     throw new AppError(400, 'Author not found')
   }
-  const result = await blog.create(payload)
-  return result
+  // Create the blog
+  const createdBlog = await blog.create(payload)
+
+  // Populate the 'author' field in the created blog
+  const populatedBlog = await blog.findById(createdBlog._id).populate('author')
+
+  return populatedBlog
 }
-const updateBlogIntroDB = async (id: string, payload: Partial<IBlog>) => {
+const updateBlogIntroDB = async (
+  id: string,
+  payload: Partial<IBlog>,
+  token: string
+) => {
+  const Author = await blog
+    .findOne({ _id: id })
+    .populate<{ author: IUser }>('author')
+  if (!Author) {
+    throw new AppError(404, 'Author not found')
+  }
+  const { email } = Author.author
+
+  const verifyAuthor = jwt.verify(
+    token,
+    config.JWT_SECRET as string
+  ) as JwtPayload
+
+  // if (verifyAuthor.email != Author.author.email) {//-
+  if (verifyAuthor.email !== email) {
+    //+
+    throw new AppError(401, 'this is not your blog')
+  }
+
   const result = await blog.findByIdAndUpdate(id, payload, {
     new: true,
     runValidators: true,
   })
   return result
 }
-const deleteBlogIntroDB = async (id: string) => {
+const deleteBlogIntroDB = async (id: string, token: string) => {
+  const Author = await blog
+    .findOne({ _id: id })
+    .populate<{ author: IUser }>('author')
+  if (!Author) {
+    throw new AppError(404, 'Author not found')
+  }
+  const { email } = Author.author
+  const verifyAuthor = jwt.verify(
+    token,
+    config.JWT_SECRET as string
+  ) as JwtPayload
+  if (verifyAuthor.email !== email) {
+    throw new AppError(401, 'this is not your blog')
+  }
   const result = await blog.findByIdAndDelete(id)
   return result
 }
@@ -50,7 +95,7 @@ const getAllBlogIntroDB = async (query: Record<string, unknown>) => {
   if (query.sortBy && query.sortOrder == 'desc') {
     sort = `-${query.sortBy as string}`
   }
-  const sortQuery = await filterQuery.sort(sort)
+  const sortQuery = await filterQuery.sort(sort).populate('author');
   return sortQuery
 }
